Adding a simple days until function
This was tested in windows gitbash so it needs to be tested on an actual linux box. The main problem was making sure i was getting a base10 number and not an octet. getting years away will involve a bit more logic but I don't know if i care enough to add it.